feat: add province distribution section
