+++ components/Sections/MainStatisticsSection.vue
